requirements.txt
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
pyyaml
joblib
shap
jupyter
papermill